Refactor state utilities into separate modules
