+++ src/components/LeadsTab.tsx
@@ -6,7 +6,8 @@ import * as yup from "yup";
 import Breadcrumbs from "./Breadcrumbs";
 import Modal from "./Modal";
 import { FixedSizeList, ListChildComponentProps } from "react-window";
-import { todayISO, saveDB, uid, fmtDate } from "../state/appState";
+import { todayISO, uid, fmtDate } from "../state/utils";
+import { saveDB } from "../state/appState";
 import type { DB, Lead, LeadStage, StaffMember, LeadFormValues } from "../types";
 
 export default function LeadsTab({ db, setDB }: { db: DB; setDB: (db: DB) => void }) {
